+++ internal/supervisor/events.go
@@ -11,10 +11,15 @@ const (
 	MemoryHigh
 	// MemoryLow indicates the process memory is at or below the threshold.
 	MemoryLow
+	// HealthOK indicates the that the health check function returned a 'nil' error
+	HealthOK
+	// HealthBad indicates the that the health check function returned an error
+	HealthBad
 )
 
 // Event is used to notify a listener of process state changes.
 type Event struct {
-	Type EventType
-	Pid  int
+	Type  EventType
+	Pid   int
+	Error error
 }