+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 UNRELEASED
 
+- Add health checks for gitaly-ruby
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/678
 - Add config option to point to languages.json
   https://gitlab.com/gitlab-org/gitaly/merge_requests/652